split train_model.py into two parts, add convert_trained_model_to_framework.py
